requirements.txt
httpx
yt-dlp
bilibili-api
msgspec
Pillow
